ci: run LSP20 <> LSP6 proxy tests in parallel separate test suite
