+++ .github/workflows/build-lint-test.yml
@@ -55,6 +55,7 @@ jobs:
             "lsp9",
             "lsp11",
             "lsp20",
+            "lsp20init",
             "universalfactory",
             "mocks",
           ]
