+++ package.json
@@ -51,7 +51,8 @@
     "test:lsp8": "hardhat test --no-compile tests/LSP8IdentifiableDigitalAsset/*.test.ts tests/LSP8IdentifiableDigitalAsset/extensions/*.test.ts",
     "test:lsp9": "hardhat test --no-compile tests/LSP9Vault/*.test.ts",
     "test:lsp11": "hardhat test --no-compile tests/LSP11BasicSocialRecovery/*.test.ts",
-    "test:lsp20": "hardhat test --no-compile tests/LSP20CallVerification/LSP6KeyManager/*.test.ts",
+    "test:lsp20": "hardhat test --no-compile tests/LSP20CallVerification/LSP6KeyManager/LSP6KeyManager.test.ts",
+    "test:lsp20init": "hardhat test --no-compile tests/LSP20CallVerification/LSP6KeyManager/LSP6KeyManagerInit.test.ts",
     "test:universalfactory": "hardhat test --no-compile tests/LSP16UniversalFactory/LSP16UniversalFactory.test.ts",
     "test:foundry": "forge test --no-match-test Skip -vvv --gas-report > gasreport.ansi",
     "build": "hardhat compile",
